# src/main/resources/apidocs.json
{
    "swagger": "2.0",
    "info": {
        "description": "Activage Data Intergration API documentation",
        "version": "0.0.1",
        "title": "Activage Data Intergration API",
        "contact": {},
        "license": {}
    },
    "host": "activage.datascienceinstitute.ie:8025",
    "basePath": "/",
    "tags": [

        {
            "name": "data-index-resource",
            "description": "Data Index Resource"
        }

    ],
    "paths": {

        "/api/createIndex": {
            "post": {
                "tags": [
                    "data-index-resource"
                ],
                "summary": "createDataIndex",
                "operationId": "createDataIndexUsingPOST",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "dataIndex",
                        "description": "dataIndex",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateIndexVM"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/DataIndex"
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            }
        },
        "/api/deleteIndex/{id}": {
            "delete": {
                "tags": [
                    "data-index-resource"
                ],
                "summary": "deleteDataIndex",
                "operationId": "deleteDataIndexUsingDELETE",
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "deprecated": false
            }
        },

        "/api/getAllIndex": {
            "get": {
                "tags": [
                    "data-index-resource"
                ],
                "summary": "getAllDataIndices",
                "operationId": "getAllDataIndicesUsingGET",
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "offset",
                        "in": "query",
                        "required": false,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number of the requested page",
                        "required": false,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "pageNumber",
                        "in": "query",
                        "required": false,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "required": false,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "paged",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "description": "Size of a page",
                        "required": false,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sorting criteria in the format: property(,asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
                        "required": false,
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi"
                    },
                    {
                        "name": "sort.sorted",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "sort.unsorted",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "unpaged",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/DataIndex"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            }
        },
        "/api/getIndex/{id}": {
            "get": {
                "tags": [
                    "data-index-resource"
                ],
                "summary": "getDataIndex",
                "operationId": "getDataIndexUsingGET",
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/DataIndex"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            }
        },

        "/api/updateIndex": {
            "put": {
                "tags": [
                    "data-index-resource"
                ],
                "summary": "updateDataIndex",
                "operationId": "updateDataIndexUsingPUT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "dataIndex",
                        "description": "dataIndex",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdateIndexOption"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/DataIndex"
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            }
        }

    },
    "definitions": {
        "CreateIndexOption": {
            "type": "object",
            "required": [
                "indexBy"
            ],
            "properties": {
                "indexBy": {
                    "type": "string",
                    "minLength": 5,
                    "maxLength": 30
                },
                "parameters": {
                    "type": "object"
                }
            },
            "title": "CreateIndexOption"
        },
        "CreateIndexVM": {
            "type": "object",
            "properties": {
                "options": {
                    "$ref": "#/definitions/CreateIndexOption"
                }
            },
            "title": "CreateIndexVM"
        },
        "DataIndex": {
            "type": "object",
            "required": [
                "indexBy"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "indexBy": {
                    "type": "string",
                    "minLength": 5,
                    "maxLength": 30
                },
                "parameters": {
                    "type": "object"
                }
            },
            "title": "DataIndex"
        },
        "UpdateIndexOption": {
            "type": "object",
            "required": [
                "indexBy",
                "indexID"
            ],
            "properties": {
                "indexBy": {
                    "type": "string",
                    "minLength": 5,
                    "maxLength": 30
                },
                "indexID": {
                    "type": "string",
                    "minLength": 5,
                    "maxLength": 30
                },
                "parameters": {
                    "type": "object"
                }
            },
            "title": "UpdateIndexOption"
        }

    }
}